refactor(cards): move the card archives to the him package
