+++ backend/src/main/java/com/brokendev/backend/card/service/CardService.java
@@ -1,23 +1,21 @@
-package com.brokendev.backend.services;
+package com.brokendev.backend.card.service;
 
 
 import com.brokendev.backend.account.domain.Account;
-import com.brokendev.backend.domain.Card;
-import com.brokendev.backend.dto.card.CardBlockResponseDTO;
-import com.brokendev.backend.dto.card.CardCreateRequestDTO;
-import com.brokendev.backend.dto.card.CardResponseDTO;
+import com.brokendev.backend.card.domain.Card;
+import com.brokendev.backend.card.dto.CardBlockResponseDTO;
+import com.brokendev.backend.card.dto.CardCreateRequestDTO;
+import com.brokendev.backend.card.dto.CardResponseDTO;
 import com.brokendev.backend.common.exceptions.UserAccountNotFoundException;
 import com.brokendev.backend.common.exceptions.CardNotFoundException;
 import com.brokendev.backend.account.domain.AccountRepository;
-import com.brokendev.backend.repositories.CardRepository;
+import com.brokendev.backend.card.domain.CardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.util.List;
-
-import static com.brokendev.backend.utils.CardUtils.*;
 
 @Service
 @RequiredArgsConstructor
